Rename package to com.farpost.groovysh
Use the canonical Groovy ecosystem abbreviation `groovysh` (the name of
the actual REPL binary in Groovy distributions, also the package of
org.apache.groovy.groovysh.Groovysh that we wrap) instead of the more
literal but redundant `groovyshell` chosen in the previous commit.

- Removes stutter in fully-qualified names: com.farpost.groovysh.GroovyShellService
  reads cleaner than com.farpost.groovyshell.GroovyShellService
- Matches the original me.bazhenov.groovysh naming convention
- Aligns with how the rest of the Groovy ecosystem refers to the shell
