+++ groovy-shell/src/main/java/com/farpost/groovysh/GroovyShellCommand.java
@@ -1,4 +1,4 @@
-package com.farpost.groovyshell;
+package com.farpost.groovysh;
 
 import groovy.lang.Binding;
 import groovy.lang.Closure;
@@ -18,7 +18,7 @@ import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import static java.util.Collections.singletonList;
-import static com.farpost.groovyshell.GroovyShellService.SHELL_KEY;
+import static com.farpost.groovysh.GroovyShellService.SHELL_KEY;
 import static org.codehaus.groovy.tools.shell.IO.Verbosity.DEBUG;
 
 class GroovyShellCommand implements Command {
