+++ groovy-shell/src/main/java/com/farpost/groovysh/SshTerminal.java
@@ -1,4 +1,4 @@
-package com.farpost.groovyshell;
+package com.farpost.groovysh;
 
 import static java.lang.Integer.parseInt;
 import static java.util.Objects.requireNonNull;
